% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.accplot.R
\name{comp.accplot}
\alias{comp.accplot}
\title{Accumulation of Occurrences Plot}
\usage{
comp.accplot(
  mods,
  nocc,
  ncells,
  xrange = c(0, ncells),
  sp.name,
  mods.names,
  alpha
)
}
\arguments{
\item{mods}{list with as much elements as models to be compared, each element must be the resulting matrix of values from the function 'accum.occ' applied to the same occurrence data (first two columns only)}

\item{nocc}{number of occurrence points}

\item{ncells}{number of cells in M}

\item{xrange}{Range of cells. Has value c(0,ncells) by default.}

\item{sp.name}{character chain with the name of the species under study}

\item{mods.names}{character vector with the names of the SDMs to be compared, must have the same length as 'mods'}

\item{alpha}{values between 0 and 1 representing the probability of the CI for the null model}
}
\value{
Plot comparing the accumulation of occurrences of each algorithm.
}
\description{
Compare the occurrence-accumulation curves \insertCite{jimenezLeavingAreaReceiving2020a}{sdmTools} of different SDM/ENMs for a single species
}
\examples{
\dontrun{
# DO NOT RUN
## Occurrence data ####
We'll only use the train data for the AOcCs
occ_train <- rgdal::readOGR('data/processed/shapefiles/train.shp')
## Environmental Data #####
envpres <- dir(path = 'data/processed/envcropped/Present/', pattern = ".tif$", full.names = T)
envpres <- raster::stack(envpres)
## Prediction Data #####
# Output from sdm::predict(), saved with terra and with full layer names retained before
# saving with names(p_occ) <- p_occ@z$fullname

pred_pres <- terra::rast(x = 'data/processed/model-build/predictions/
                         predictions.present-all-algorithms.tif')
# remember to read with terra:: to retain layer names
pred_pres <- raster::stack(pred_pres) # and then turn them into RasterStacks

## AOcCs of models #####
# Prepare the data to pass it onto the function:
occ.pnts <- sdmTools::sdm_to_occ.pnts(env = envpres,
                           occ = occ_train,
                           algorithms = c('svm', 'maxent', 'bioclim', 'domain'),
                           # list of algorithms used in sdm::sdm()
                           predict_object = pred_pres)

output.mod <- sdmTools::sdm_to_output.mod(env = envpres,
                                algorithms = c('svm', 'maxent', 'bioclim', 'domain'),
                                predict_object = pred_pres)
# The accum.occ function is annoying and creates three different plots in three different devices.
# To save them we'll use dev.print() in the reverse order in which they appear. I could automatize
# this part of the script but I don't have time for that, I'm sorry dear reader.

dev.new() ; dev.control('enable') ; dev.off() # Just enabling device control

# WARNING: Before running this next section of the script, make sure nothing
# is plotted and no devices are open with dev.list()
aocc_svm <- accum.occ(sp.name = 'Charinus',
                     output.mod = output.mod[[1]],
                     occ.pnts = occ.pnts[[1]],
                     null.mod = "hypergeom",
                     conlev = 0.05, bios = 0)

# To save the plots, we have to do it in the reverse order as they were plotted:
dev.print(png, file = "figs/aocc/svm_aocc_aocc.png",
          width=8, height=8, units="in", res=300);dev.off()

dev.print(png, file = "figs/aocc/svm_aocc_map.png",
          width=8, height=8, units="in", res=300);dev.off()

dev.print(png, file = "figs/aocc/svm_aocc_env.png",
          width=8, height=8, units="in", res=300);dev.off()

# Do this for each algorithm and when you have all of them, compare with

aocc_list <- list(aocc_svm, aocc_maxent, aocc_bioclim, aocc_domain)

model_comp <- comp.accplot(mods = aocc_list,
                          nocc = length(occ_train),
                          ncells = raster::ncell(envpres),
                          sp.name = 'Charinus',
                          mods.names = c('svm', 'maxent', 'bioclim', 'domain'), alpha = 0.05)
}
}
\references{
\insertRef{jimenezLeavingAreaReceiving2020a}{sdmTools}
}
\author{
Laura Jiménez
}
