% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congruence}
\alias{congruence}
\title{Congruence Map}
\format{
A RasterStack of global-scale bioclimatic congruence map to analyse environmental mismatches between recently updated bioclimatic databases \href{www.worldclim.org/version2}{WorldClim v. 2},MERRAclim (\insertCite{c.vegaMERRAclimHighresolutionGlobal2017}{sdmLit} , \insertCite{c.vegaMERRAclimHighresolutionGlobal2017a}{sdmLit}), \href{www.chelsa-climate.org}{CHELSA}, and ocean databases \href{www.marspec.org}{MARSPEC} and \href{www.bio-oracle.org}{Bio-ORACLE v. 2.0}:
\describe{
\item{\code{class}}{RasterLayer}
\item{\code{dimensions}}{1080, 2160, 2332800  (nrow, ncol, ncell)}
\item{\code{resolution}}{0.1666667, 0.1666667  (x, y)}
\item{\code{extent}}{180, 180, -90, 90  (xmin, xmax, ymin, ymax)}
\item{\code{crs}}{NA}
\item{\code{source}}{inst/extdata/congruence.grd}
\item{\code{names}}{congruence}
\item{\code{values}}{0, 1  (min, max)}
}
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.6kv7k29/}
}
\usage{
congruence
}
\description{
Map of bioclimatic congruence (congruence.grd) from \insertCite{morales-barberoInputMattersMatter2019a}{sdmLit}.
}
\references{
\insertRef{morales-barberoInputMattersMatter2019a}{sdmLit}

\insertRef{c.vegaMERRAclimHighresolutionGlobal2017}{sdmLit}

\insertRef{c.vegaMERRAclimHighresolutionGlobal2017a}{sdmLit}
}
\keyword{datasets}
