% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_to_boyceIndex.R
\name{sdm_to_boyceIndex}
\alias{sdm_to_boyceIndex}
\title{Boyce Index from sdm::predict() Output}
\usage{
sdm_to_boyceIndex(preds, layer, occurrence, ...)
}
\arguments{
\item{preds}{RasterStack. Output from sdm::predict().}

\item{layer}{String, Function or Interger. If string, name of the layer to be used exactly as found on the preds layer. If an Interger, the number of the layer to be used. Can be a function e.g. mean(2:5), which will average the layers and then perform the analyses.}

\item{occurrence}{Either a SpatialPointsDataFrame or a data.frame. If a data.frame, x and y (lat and long) coordinates of the occurrence records as two separate collumns. Usually used with the test data, instead of the train data.}

\item{...}{additional arguments to be passed onto ecospat.boyce()}
}
\value{
A list, and by default PEplot is plotted to the device. List contains three items: F.ratio, cor and HS. For more information please refer to \insertCite{hirzelEvaluatingAbilityHabitat2006}{sdmLit} or to the ecospat.boyce() help (linked below).
}
\description{
This function simply converts the output from the sdm package to the Boyce Index \insertCite{hirzelEvaluatingAbilityHabitat2006}{sdmLit} using the function from the ecospat package.
}
\examples{
occ_test <- rgdal::readOGR('data/processed/shapefiles/test.shp')
occ <- as.data.frame(occ_test)
pred_pres <- terra::rast(x = 'data/processed/model-build/predictions/predictions.present-terraStd.tif') # remember to read with terra:: to retain layer names
pred_pres <- raster::stack(pred_pres) # and then turn them into RasterStacks

teste <- sdm_to_boyceIndex(pred_pres, mean(2:5), occ[,2:3])

}
\references{
\insertRef{hirzelEvaluatingAbilityHabitat2006}{sdmLit}
}
\seealso{
\code{\link[ecospat:ecospat.boyce]{ecospat::ecospat.boyce()}} which this function wraps.
}
