% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_to_freqEnsemble.R
\name{sdm_to_freqEnsemble}
\alias{sdm_to_freqEnsemble}
\title{Frequency Ensemble from sdm::sdm() and sdm::predict()}
\usage{
sdm_to_freqEnsemble(model, preds, species, threshold)
}
\arguments{
\item{model}{Model object as output from sdm() function or written/read with sdm::read.sdm().}

\item{preds}{Prediction Object Output, with two observations: first, for it to have species names and
algorithm names in the layer names, you have to run  \code{names(p_occ) <- p_occ@z$fullname } and second,
you have to save (and later read it) it with the terra package to retain layer names
\code{p_occ_stack <- raster::stack(p_occ) ; }
\code{p_occ_spatraster <- terra::rast(p_occ) ; }
\code{names(p_occ_spatraster) <- names(p_occ_stack) ; }
\verb{terra::writeRaster(x = p_occ_spatraster, }
\verb{                   filename = 'data/processed/final-model-build/predictions/predictions.present-terraStd.tif', overwrite = TRUE)}}

\item{species}{Character. Names of the species as they appear in layer names.}

\item{threshold}{Interger. Which parameter to use as threshold for binarizing the predictions. The possible value can be between 1 to 10 for "sp=se", "max(se+sp)", "min(cost)", "minROCdist", "max(kappa)", "max(ppv+npv)", "ppv=npv", "max(NMI)", "max(ccr)", "prevalence" criteria, respectively. For more information, see \code{\link[sdm:evaluates]{sdm::getEvaluation()}}.}
}
\value{
A RasterStack with one layer, ensembled all models present in 'model' and 'preds' by frequency of presences as of threshold max(sp+se). See  \insertCite{sobral-souzaSpeciesExtinctionRisk2015}{sdmTools} or \insertCite{dasilveiraFutureClimateChange2021}{sdmTools} for details.
}
\description{
This function reproduces the methods from Sobral-Souza and Prasniewski, in which ensembling by the frequency among model predictions is made. For more information please consult the literature in \insertCite{sobral-souzaSpeciesExtinctionRisk2015}{sdmTools} or \insertCite{dasilveiraFutureClimateChange2021}{sdmTools}.
}
\examples{
\dontrun{
## DO NOT RUN
library(raster);library(sdm);library(dplyr);library(ggplot2)
model <- sdm::read.sdm("data/processed/model-build/model-object/model.sdm")
pocc <- terra::rast('data/processed/model-build/predictions/predictions.present-terraStd.tif')
preds <- raster::stack(pocc)
teste <- sdm_to_freqEnsemble(model, preds, c('charinus'), 1) }
}
\references{
\insertRef{sobral-souzaSpeciesExtinctionRisk2015}{sdmTools}

\insertRef{dasilveiraFutureClimateChange2021}{sdmTools}
}
\seealso{
\code{\link[sdm:evaluates]{sdm::getEvaluation()}}
}
