% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_to_output.mods.R
\name{sdm_to_output.mods}
\alias{sdm_to_output.mods}
\title{sdm to output.mod}
\usage{
sdm_to_output.mods(env, algorithms, predict_object, long, lat)
}
\arguments{
\item{env}{RasterStack used as newdata= in sdm::predict() function}

\item{algorithms}{Vector of strings. Names of the methods passed as parameters in the sdm function, passed here identically. Order is important. If any of the algorithms failed you should not pass it onto this function. Example: algorithms = c('svm', 'maxent', 'rf', 'brt', 'domain')}

\item{predict_object}{RasterStack. Object obtained from sdm::predict() function. Can be the object as expelled from the function or the saved output loaded as RasterStack}

\item{long}{String. How the long variable is called in your output from sdm::predict(). Example: long = 'x'}

\item{lat}{String. How the long variable is called in your output from sdm::predict(). Example: lat = 'y'}
}
\value{
Returns a list of dataframes as long as the number of algorithms passed in 'algorithms ='. Each dataframe is in the output.mod format required by the accum.occ function and contains the information of one algorithm.
}
\description{
This function takes the output from the sdm::predict() function, and environmental data used in predict() and makes a list of data.frames, one data.frame for each algorithm, in the output.mod format required by the accum.occ() function from \insertCite{jimenezLeavingAreaReceiving2020a}{sdmLit}.
}
\examples{
\dontrun{ # DO NOT RUN
envpres <- dir(path = 'path/to/cropped/env/layers', pattern = ".tif$", full.names = T)
envpres <- raster::stack(envpres)

occ_train <- rgdal::readOGR('data/processed/shapefiles/train.shp')

d_occ <- sdmData(formula = charinus~., train=occ_train,
                 predictors = envpres, bg = 18,
                 method = 'eRandom') # Create sdmData object

m_occ <- sdm(formula = charinus~.,data = d_occ,
             methods = c('svm', 'maxent', 'bioclim', 'domain'),
                          replication = c('bootstrapping'), n=5)
p_occ <- predict(m_occ, newdata = envpres, # new data is the environment in which we want to predict
                 filename = 'output/from/predict/function.tif',
                   mean = T, prj = T, overwrite = T, nc = 4)

out.preds <- sdm_to_output.mods(env = envpres, long = 'x',lat = 'y',
                                algorithms = c('svm', 'maxent','bioclim','domain'),
                                predict_object = p_occ)
          }
}
\references{
\insertRef{jimenezLeavingAreaReceiving2020a}{sdmLit}
}
\author{
João Frederico Berner
}
